requirements.txt
autopep8==2.3.1
backports.tarfile==1.2.0
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
importlib-metadata==8.0.0
importlib-resources==6.4.0
isort==5.13.2
jaraco.collections==5.1.0
packaging==24.1
platformdirs==4.2.2
python-decouple==3.8
tomli==2.0.1
